The new, refactored hotness.
